pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "gpu-usage-menubar"
version = "1.0.0"
description = "A macOS menu bar app for monitoring remote GPU utilization"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Vijay Daultani"}
]
dependencies = [
    "pyobjc-framework-Cocoa>=10.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
gpu-usage-menubar = "gpu_usage_menubar.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/gpu_usage_menubar --cov-report=term-missing"
